user/account_validation/requirements.txt
falcon
gunicorn
pymongo
pyjwt
falcon-cors
rq